chore: extract orderbook-commons to be able to resuse types
