+++ crates/orderbook-client/Cargo.toml
@@ -9,8 +9,9 @@ anyhow = "1"
 async-stream = "0.3"
 futures = "0.3"
 hex = "0.4.3"
+orderbook-commons = { path = "../orderbook-commons" }
 ring = "0.16"
-secp256k1-zkp = { version = "0.7.0", features = ["global-context"] }
+secp256k1 = { version = "0.24.3", features = ["global-context", "serde"] }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 sha2 = { version = "0.10", default-features = false }
