+++ crates/orderbook-commons/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "orderbook-commons"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1"
+secp256k1 = { version = "0.24.3", features = ["serde"] }
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+sha2 = { version = "0.10", default-features = false }
